Move business logic into ToyClient class
